+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes will be documented here.
 
+## [v2.1.0] - 2025-09-12
+
+- don't attempt to push if there is already a pending packfile update proposal
+- set GITOPIA_ENV to prod by default in Makefile
+
 ## [v2.0.1] - 2025-09-08
 
 - handle the cases when there is only ref update on chain